feat(source): single source for proposals as strategies
